+++ python/glow/wgr/tests/test_block_variants_and_samples.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from glow import glow
-from glow.wgr import functions
+from glow.wgr import wgr_functions
 import pytest
 from pyspark.sql import Row
 from pyspark.sql.functions import expr
@@ -33,10 +33,10 @@ def test_block_variants_and_samples(spark):
         .load("test-data/combined.chr20_18210071_18210093.g.vcf") \
         .withColumn("values", expr("genotype_states(genotypes)"))
     sample_ids = ["HG00096", "HG00268", "NA19625"]
-    block_gt, index_map = functions.block_variants_and_samples(variant_df,
-                                                               sample_ids,
-                                                               variants_per_block=10,
-                                                               sample_block_count=2)
+    block_gt, index_map = wgr_functions.block_variants_and_samples(variant_df,
+                                                                   sample_ids,
+                                                                   variants_per_block=10,
+                                                                   sample_block_count=2)
     expected_block_gt = glow.transform("block_variants_and_samples",
                                        variant_df,
                                        variants_per_block=10,
@@ -49,29 +49,29 @@ def test_missing_values(spark):
     variant_df = spark.read.format("vcf").load("test-data/combined.chr20_18210071_18210093.g.vcf")
     sample_ids = ["HG00096", "HG00268", "NA19625"]
     with pytest.raises(AnalysisException):
-        functions.block_variants_and_samples(variant_df,
-                                             sample_ids,
-                                             variants_per_block=10,
-                                             sample_block_count=2)
+        wgr_functions.block_variants_and_samples(variant_df,
+                                                 sample_ids,
+                                                 variants_per_block=10,
+                                                 sample_block_count=2)
 
 
 def test_no_values(spark):
     variant_df = spark.createDataFrame([__construct_row([0, 1])]).limit(0)
     sample_ids = ["a", "b"]
     with pytest.raises(Exception):
-        functions.block_variants_and_samples(variant_df,
-                                             sample_ids,
-                                             variants_per_block=10,
-                                             sample_block_count=2)
+        wgr_functions.block_variants_and_samples(variant_df,
+                                                 sample_ids,
+                                                 variants_per_block=10,
+                                                 sample_block_count=2)
 
 
 def test_inconsistent_num_values(spark):
     variant_df = spark.createDataFrame([__construct_row([0, 1, 1]), __construct_row([1, 2])])
     sample_ids = ["a", "b", "c"]
-    block_df, index_map = functions.block_variants_and_samples(variant_df,
-                                                               sample_ids,
-                                                               variants_per_block=10,
-                                                               sample_block_count=2)
+    block_df, index_map = wgr_functions.block_variants_and_samples(variant_df,
+                                                                   sample_ids,
+                                                                   variants_per_block=10,
+                                                                   sample_block_count=2)
     with pytest.raises(Exception):
         block_df.collect()  # Number of values is checked lazily within the transformer
 
@@ -80,27 +80,27 @@ def test_mismatch_num_values_sample_ids(spark):
     variant_df = spark.createDataFrame([__construct_row([0, 1]), __construct_row([1, 1])])
     sample_ids = ["a", "b", "c"]
     with pytest.raises(Exception):
-        functions.block_variants_and_samples(variant_df,
-                                             sample_ids,
-                                             variants_per_block=10,
-                                             sample_block_count=2)
+        wgr_functions.block_variants_and_samples(variant_df,
+                                                 sample_ids,
+                                                 variants_per_block=10,
+                                                 sample_block_count=2)
 
 
 def test_missing_sample_ids(spark):
     variant_df = spark.createDataFrame([__construct_row([0, 1]), __construct_row([1, 1])])
     sample_ids = ["a", ""]
     with pytest.raises(Exception):
-        functions.block_variants_and_samples(variant_df,
-                                             sample_ids,
-                                             variants_per_block=10,
-                                             sample_block_count=2)
+        wgr_functions.block_variants_and_samples(variant_df,
+                                                 sample_ids,
+                                                 variants_per_block=10,
+                                                 sample_block_count=2)
 
 
 def test_duplicated_sample_ids(spark):
     variant_df = spark.createDataFrame([__construct_row([0, 1]), __construct_row([1, 1])])
     sample_ids = ["a", "a"]
     with pytest.raises(Exception):
-        functions.block_variants_and_samples(variant_df,
-                                             sample_ids,
-                                             variants_per_block=10,
-                                             sample_block_count=2)
+        wgr_functions.block_variants_and_samples(variant_df,
+                                                 sample_ids,
+                                                 variants_per_block=10,
+                                                 sample_block_count=2)
