+++ python/glow/wgr/tests/test_functions.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from glow.wgr.linear_model.functions import *
+from glow.wgr.model_functions import *
 import math
 import numpy as np
 import pandas as pd
