+++ python/glow/wgr/tests/test_sample_id_extraction.py
@@ -15,7 +15,7 @@
 import pytest
 from pyspark.sql import Row
 from pyspark.sql.utils import AnalysisException
-from glow.wgr import functions
+from glow.wgr import wgr_functions
 
 
 def __construct_row(sample_id_1, sample_id_2):
@@ -28,7 +28,7 @@ def __construct_row(sample_id_1, sample_id_2):
 
 def test_get_sample_ids(spark):
     df = spark.read.format("vcf").load("test-data/combined.chr20_18210071_18210093.g.vcf")
-    sample_ids = functions.get_sample_ids(df)
+    sample_ids = wgr_functions.get_sample_ids(df)
     assert (sample_ids == ["HG00096", "HG00268", "NA19625"])
 
 
@@ -36,28 +36,28 @@ def test_missing_sample_id_field(spark):
     df = spark.read.format("vcf").option("includeSampleIds", "false") \
         .load("test-data/combined.chr20_18210071_18210093.g.vcf")
     with pytest.raises(AnalysisException):
-        functions.get_sample_ids(df)
+        wgr_functions.get_sample_ids(df)
 
 
 def test_inconsistent_sample_ids(spark):
     df = spark.createDataFrame([__construct_row("a", "b"), __construct_row("a", "c")])
     with pytest.raises(Exception):
-        functions.get_sample_ids(df)
+        wgr_functions.get_sample_ids(df)
 
 
 def test_incorrectly_typed_sample_ids(spark):
     df = spark.createDataFrame([__construct_row(1, 2)])
     with pytest.raises(Exception):
-        functions.get_sample_ids(df)
+        wgr_functions.get_sample_ids(df)
 
 
 def test_empty_sample_ids(spark):
     df = spark.createDataFrame([__construct_row("a", "")])
     with pytest.raises(Exception):
-        functions.get_sample_ids(df)
+        wgr_functions.get_sample_ids(df)
 
 
 def test_duplicated_sample_ids(spark):
     df = spark.createDataFrame([__construct_row("a", "a")])
     with pytest.raises(Exception):
-        functions.get_sample_ids(df)
+        wgr_functions.get_sample_ids(df)
