+++ python/glow/wgr/wgr_functions.py
